+++ src/utils/query-keys.ts
@@ -1,0 +1,4 @@
+export const QUERY_KEY_PATIENT_CONDITIONS = "patientConditions";
+export const QUERY_KEY_OTHER_PROVIDER_CONDITIONS = "otherProviderConditions";
+export const QUERY_KEY_CONDITION_HISTORY = "conditionHistory";
+export const QUERY_KEY_PATIENT = "patient";